% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestTool.atleast.R
\name{BestTool.atleast}
\alias{BestTool.atleast}
\title{BestTool.atleast function}
\usage{
BestTool.atleast(BestPat)
}
\arguments{
\item{BestPat}{the table with summarized data generated with SumSingleTools() function}
}
\description{
This function computes confusion matrices and accuracy values for all variants of "at least n tools".
}
\examples{
BP.atl <- BestTool.atleast(BestPat = BestPat2)
BP.atl
}
\keyword{accuracy}
\keyword{confiusion}
\keyword{lncRNA}
\keyword{matrix}
\keyword{precision}
