% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestTool.comb.R
\name{BestTool.comb}
\alias{BestTool.comb}
\title{BestTool.comb function}
\usage{
BestTool.comb(BestPat, selectComb)
}
\arguments{
\item{BestPat}{the table with summarized data generated with SumSingleTools() function}
}
\description{
This function computes confusion matrices and accuracy values for all combinations of all selected tools.
}
\examples{
BP.cmb <- BestTool.comb(BestPat = BestPat3)
BP.cmb
}
\keyword{accuracy}
\keyword{confiusion}
\keyword{lncRNA}
\keyword{matrix}
\keyword{precision}
