% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestTool.R
\name{BestTool}
\alias{BestTool}
\title{BestTool function}
\usage{
BestTool(BestPat, tools)
}
\arguments{
\item{BestPat}{the table with summarized data generated with SumSingleTools() function}

\item{tools}{the list of names of tools to compute the accuracy}
}
\description{
This function computes confusion matrix and accuracy for single methods .
}
\examples{
selectedTools <- c("CPC2", "PLEK", "FEELnc", "CPAT", "CNCI", "LncFinder")
BP.cpt <- BestTool(BestPat = BestPat1, selectedTools)
BP.cpt
}
\keyword{lncRNA}
\keyword{test}
\keyword{training}
