% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cisInter.R
\name{cisInter}
\alias{cisInter}
\title{cisInter function
This function reads several outputs from coding potential programs and join all tesults in table, where 0 is coding and 1 - noncoding}
\usage{
cisInter(
  lncRNAs = NULL,
  mRNAs = NULL,
  FEELnc.classes,
  is.best = T,
  lncRNA.level = "transcript",
  mRNA.level = "gene",
  max.dist = 1e+05
)
}
\arguments{
\item{lncRNAs}{the list of lncRNA IDs (genes or transcripts)}

\item{mRNAs}{the list of mRNA IDs (genes or transcripts)}

\item{FEELnc.classes}{the "classes" output file from FEELnc}

\item{is.best}{filtering by the first collumn "is.best" (see https://github.com/tderrien/FEELnc)}

\item{lncRNA.level}{choose the "gene" or "transcript" level for lncRNA filtering}

\item{mRNA.level}{choose the "gene" or "transcript" level for mRNA filtering}

\item{max.dist}{the distance for cis-relation (default value is 100 000)}
}
\description{
cisInter function
This function reads several outputs from coding potential programs and join all tesults in table, where 0 is coding and 1 - noncoding
}
\examples{
cisInter()

}
\keyword{cis}
\keyword{interactions}
\keyword{lncRNA}
