% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodPot.R
\name{CodPot2tbl}
\alias{CodPot2tbl}
\title{A CodPot2tbl function}
\usage{
CodPot2tbl(
  CPC2_outfile = NULL,
  PLEK_outfile = NULL,
  FEELnc_outfile = NULL,
  CPAT_outfile = NULL,
  CPAT_cutoff = 0.364,
  CNCI_outfile = NULL,
  LncFinder_outfile = NULL,
  lncRNA_Mdeep_outfile = NULL
)
}
\arguments{
\item{CPC2_outfile}{the output file from the CPC2 tool}

\item{PLEK_outfile}{the output file from the PLEK tool}

\item{FEELnc_outfile}{the output file from the FEELnc tool}

\item{CPAT_outfile}{the output file from the CPAT tool}

\item{CPAT_cutoff}{the cutoff value for CPAT output}

\item{CNCI_outfile}{the output file from the CNCI tool}

\item{LncFinder_outfile}{the output file from the LncFinder tool}

\item{lncRNA_Mdeep_outfile}{the output file from the lncRNA_Mdeep tool}
}
\description{
This function reads several outputs from coding potential programs and join all tesults in table, where 0 is coding and 1 - noncoding
}
\examples{
CodPot2tbl()

}
\keyword{coding}
\keyword{lncRNA}
\keyword{potential}
