% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10Fold_CrossValidation_modified.R
\name{CPATcutoff}
\alias{CPATcutoff}
\title{A CPATcutoff function}
\usage{
CPATcutoff(CPATout.xls)
}
\arguments{
\item{CPATout.xls}{is the one of the CPAT output files containing coding probability computet for training dataset}
}
\description{
This function reads CPAT output and estimates the cutoff value
}
\examples{
CPATcutoff()

}
\keyword{CPAT}
\keyword{cutoff}
\keyword{lncRNA}
