% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/express.R
\name{ExpMeans}
\alias{ExpMeans}
\title{ExpMeans function}
\usage{
ExpMeans(ExprTab)
}
\arguments{
\item{ExprTab}{is the table with expression valules}
}
\description{
This function allows you to calculate the mean values of the expression.
}
\examples{
ExpMeans()
}
\keyword{expression}
\keyword{lncRNA}
