% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/express.R
\name{ExpSums}
\alias{ExpSums}
\title{ExpSums function}
\usage{
ExpSums(ExprTab)
}
\arguments{
\item{ExprTab}{is the table with expression valules}
}
\description{
This function allows you to sum expression values.
}
\examples{
ExprTab()
}
\keyword{expression}
\keyword{lncRNA}
