% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpFun.R
\name{\%!in\%}
\alias{\%!in\%}
\title{%!in% function}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{vector of values to be filtered}

\item{y}{vector of values to eliminate}
}
\description{
This function is negation of %in%.
}
\examples{
vector_A \%!in\% vector_B

}
\keyword{negate}
