% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpFun.R
\name{h.t}
\alias{h.t}
\title{h.t function}
\usage{
h.t(table, n = 3)
}
\arguments{
\item{table}{table}

\item{n}{number of rows to display}
}
\description{
Display head and tail together.
}
\examples{
h.t(table)

}
\keyword{head}
\keyword{table}
\keyword{tail}
