% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lncRna-package.R
\docType{package}
\name{lncRna-package}
\alias{lncRna}
\alias{lncRna-package}
\title{lncRna: A Comprehensive Workflow for Long Non-coding RNA Identification and Functional Analysis}
\description{
Provides a complete workflow for the identification, analysis, and functional annotation of long non-coding RNAs (lncRNAs) from RNA-Seq data. The package includes functions for filtering transcripts from GTF files, evaluating the performance of multiple coding potential prediction tools (e.g., CPC2, PLEK, CPAT), and summarizing their agreement. It enables systematic performance analysis of individual tools, "at least N" tool consensus, and all possible tool combinations. Functional analysis is supported through the identification of potential cis- and trans-acting interactions with protein-coding genes, followed by enrichment analysis. Results can be visualized using a variety of plots, including radar plots, clock plots, and interactive Sankey diagrams.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/prodakt/lncRna}
  \item Report bugs at \url{https://github.com/prodakt/lncRna/issues}
}

}
\author{
\strong{Maintainer}: Jan Pawel Jastrzebski \email{bioinformatyka@gmail.com} (\href{https://orcid.org/0000-0001-8699-7742}{ORCID})

Authors:
\itemize{
  \item Damian Czopek \email{dcwmpl@gmail.com} (\href{https://orcid.org/0009-0005-3471-4866}{ORCID}) [contributor]
}

Other contributors:
\itemize{
  \item Mariusz Jankowski \email{ma.jankowski99@gmail.com} (\href{https://orcid.org/0009-0000-7872-4023}{ORCID}) [contributor]
  \item Monika Gawronska \email{gawronska572@gmail.com} (\href{https://orcid.org/0009-0001-2677-6371}{ORCID}) [contributor]
  \item Wiktor Babis \email{wiktorbabis@gmail.com} (\href{https://orcid.org/0009-0006-3648-3413}{ORCID}) [contributor]
  \item Stefano Pascarella \email{stefano.pascarella@uniroma1.it} (\href{https://orcid.org/0000-0002-6822-4022}{ORCID}) [contributor]
}

}
\keyword{internal}
