% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.R
\name{load_packages}
\alias{load_packages}
\title{Load R packages, with options to include default, CRAN, and GitHub packages.}
\usage{
load_packages(additional_packages = NULL, additional_github_packages = NULL)
}
\arguments{
\item{additional_packages}{A character vector of package names to install and load from CRAN, in addition to the default packages. Defaults to \code{NULL}.}

\item{additional_github_packages}{A character vector of GitHub package paths in the format "user/repo", to install and load, in addition to the default GitHub package ('lncRna'). Defaults to \code{NULL}.}
}
\description{
This function loads a set of default R packages, and optionally installs and loads additional packages from CRAN and GitHub.
It handles package installation and loading, providing informative messages about the process and any errors encountered.
}
\examples{
# Example 1: Load default packages only
# load_packages()

# Example 2: Load default CRAN packages PLUS 'dplyr' and 'tidyr'
# load_packages(additional_packages = c("dplyr", "tidyr"))

# Example 3: Load default packages PLUS additional GitHub packages
# load_packages(additional_github_packages = c("your_github_user/your_github_repo_package1", "another_github_user/another_package"))
# Note: For GitHub packages, you need to provide the full repo path "user/repo"

# Example 4: Load default packages PLUS additional CRAN and GitHub packages
# load_packages(additional_packages = c("dplyr", "randomForest"), additional_github_packages = c("your_github_user/your_github_repo_package1"))

# Example 5: Load default packages PLUS attempt to load a non-existent package from GitHub (error handling demo)
# load_packages(additional_github_packages = c("nonexistent_github_user/nonexistent_package"))

# Example 6: Load default packages PLUS valid and invalid GitHub packages
# load_packages(additional_github_packages = c("prodakt/lncRna", "nonexistent_github_user/nonexistent_package")) # Note: "prodakt/lncRna" is already in default, but it will handle it correctly

}
