% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateCodPot.R
\name{plotVennCodPot}
\alias{plotVennCodPot}
\title{Create Venn Diagram from Coding Potential Results}
\usage{
plotVennCodPot(
  codPot,
  selection = NULL,
  vennColors = grDevices::palette.colors(n = 9, palette = "Okabe-Ito")
)
}
\arguments{
\item{codPot}{A list object generated by \code{aggregateCodPot()}.}

\item{selection}{An optional logical vector indicating which tools to include
(TRUE for include, FALSE for exclude). If \code{NULL}, all available tools are used.}

\item{vennColors}{A vector of colors for the Venn diagram segments.}
}
\value{
Invisibly returns the result of \code{venn::venn}. The primary effect is
plotting the diagram to the active graphics device.
}
\description{
Generates a Venn diagram of noncoding predictions from multiple tools, based
on the output of \code{aggregateCodPot}. This function requires the 'venn' package,
which should be declared in the \code{Suggests} field of the DESCRIPTION file.
}
\examples{
  # Example with all tools

  mockCodPot <- list(
  seqIDs = c("tx1", "tx2", "tx3", "tx4"),
  tools = list(
    CPC2 = c(0, 1, 1, 0),
    PLEK = c(0, 1, 0, 1)
  )
)
if (requireNamespace("venn", quietly = TRUE)) {
  plotVennCodPot(codPot = mockCodPot)
}
}
