% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codPot.R
\name{read.CNCI}
\alias{read.CNCI}
\title{A read.CNCI function}
\usage{
read.CNCI(CNCI_outfile)
}
\arguments{
\item{CNCI_outfile}{is the CNCI output file localization including filename}
}
\description{
This function reads CNCI output and list all noncoding transcripts IDs
}
\examples{
read.CNCI()

}
\keyword{CNCI}
\keyword{lncRNA}
