% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodPot.R
\name{read.CPAT}
\alias{read.CPAT}
\title{A read.CPAT function}
\usage{
read.CPAT(CPAT_outfile, CPAT_cutoff = 0.364)
}
\arguments{
\item{CPAT_outfile}{is the CPAT output file (table containing 'lable' collumn inticating coding = 1 and noncoding = 0 transcripts) localization including filename}

\item{CPAT_cutoff}{is the cutoff value predefined (https://cpat.readthedocs.io/en/latest/#how-to-choose-cutoff) or computed i.e. using 'CPATcutoff()' function.
The default value is equal to predefined for human 0.364}
}
\description{
This function reads CPAT output and list all noncoding transcripts IDs
}
\examples{
read.CPAT()

}
\keyword{CPAT}
\keyword{lncRNA}
