% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodPot.R
\name{read.CPC2}
\alias{read.CPC2}
\title{A read.CPC2 function}
\usage{
read.CPC2(CPC2_outfile)
}
\arguments{
\item{CPC2_outfile}{is the CPC2 output file localization including filename}
}
\description{
This function reads CPC2 output and list all noncoding transcripts IDs
}
\examples{
read.CPC2()

}
\keyword{CPC2}
\keyword{lncRNA}
