% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codPot.R
\name{read.FEELnc}
\alias{read.FEELnc}
\title{A read.FEELnc function}
\usage{
read.FEELnc(FEELnc_outfile)
}
\arguments{
\item{FEELnc_outfile}{is the FEELnc output file (table containing 'lable' collumn inticating coding = 1 and noncoding = 0 transcripts) localization including filename}
}
\description{
This function reads FEELnc output and list all noncoding transcripts IDs
}
\examples{
read.FEELnc()

}
\keyword{FEELnc}
\keyword{lncRNA}
