% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codPot.R
\name{read.LncFinder}
\alias{read.LncFinder}
\title{A read.LncFinder function}
\usage{
read.LncFinder(LncFinder_outfile)
}
\arguments{
\item{LncFinder_outfile}{is the LncFinder output file localization including filename}
}
\description{
This function reads LncFinder output and list all noncoding transcripts IDs
}
\examples{
read.LncFinder()

}
\keyword{LncFinder}
\keyword{lncRNA}
