% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codPot.R
\name{read.lncRNA_Mdeep}
\alias{read.lncRNA_Mdeep}
\title{A read.lncRNA_Mdeep function}
\usage{
read.lncRNA_Mdeep(lncRNA_Mdeep_outfile)
}
\arguments{
\item{lncRNA_Mdeep_outfile}{is the lncRNA_Mdeep output file localization including filename}
}
\description{
This function reads lncRNA_Mdeep output and list all noncoding transcripts IDs
}
\examples{
read.lncRNA_Mdeep()

}
\keyword{lncRNA}
\keyword{lncRNA_Mdeep}
