% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pfam.R
\name{read.pfam}
\alias{read.pfam}
\title{A read.pfam function}
\usage{
read.pfam(pfam_outfile, eval_cutoff = 0.01)
}
\arguments{
\item{pfam_outfile}{is the hmmscan pfam output file localization including filename}

\item{eval_cutoff}{is the e-value cutoff}
}
\description{
This function reads hmmscan pfam output and list all noncoding transcripts IDs
}
\examples{
read.pfam()

}
\keyword{lncRNA}
\keyword{pfam}
