% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodPot.R
\name{read.PLEK}
\alias{read.PLEK}
\title{A read.PLEK function}
\usage{
read.PLEK(PLEK_outfile)
}
\arguments{
\item{PLEK_outfile}{is the PLEK output file localization including filename}
}
\description{
This function reads PLEK output and list all noncoding transcripts IDs
}
\examples{
read.PLEK()

}
\keyword{PLEK}
\keyword{lncRNA}
