% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rfam.R
\name{read.rfam}
\alias{read.rfam}
\title{A read.rfam function}
\usage{
read.rfam(rfam_outfile)
}
\arguments{
\item{rfam_outfile}{is the hmmscan pfam output file localization including filename}

\item{eval_cutoff}{is the e-value cutoff}
}
\description{
This function reads hmmscan rfam output and list all noncoding transcripts IDs
}
\examples{
read.rfam()

}
\keyword{lncRNA}
\keyword{rfam}
