% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodPotV2.R
\name{readCnci}
\alias{readCnci}
\title{Read CNCI Output}
\usage{
readCnci(cnciOutfile)
}
\arguments{
\item{cnciOutfile}{Path to the CNCI output file.}
}
\value{
A named character vector of noncoding transcript IDs.
}
\description{
Read CNCI Output
}
\examples{
# Create a temporary file
tempCnciFile <- tempfile()
write.table(
  data.frame(transcript_ID=c("id_x", "id_y"), index=c("noncoding", "coding")),
  tempCnciFile, row.names=FALSE, quote=FALSE, sep="\t"
)
# Run and print
noncodingCnci <- readCnci(tempCnciFile)
print(noncodingCnci)
# Clean up
unlink(tempCnciFile)
}
\keyword{internal}
