% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodPotV2.R
\name{readCpc2}
\alias{readCpc2}
\title{Read CPC2 Output}
\usage{
readCpc2(cpc2Outfile)
}
\arguments{
\item{cpc2Outfile}{Path to the CPC2 output file.}
}
\value{
A named character vector of noncoding transcript IDs.
}
\description{
Read CPC2 Output
}
\examples{
# Create a temporary file for the example
tempCpc2File <- tempfile()
write.table(
  data.frame(V1=c("ID1", "ID2", "ID3"), V8=c("noncoding", "coding", "noncoding")),
  tempCpc2File, row.names=FALSE, col.names=FALSE, quote=FALSE
)
# Run the reader function
noncodingCpc2 <- readCpc2(tempCpc2File)
print(noncodingCpc2)
# Clean up
unlink(tempCpc2File)
}
\keyword{internal}
