% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodPotV2.R
\name{readFeelnc}
\alias{readFeelnc}
\title{Read FEELnc Output}
\usage{
readFeelnc(feelncOutfile)
}
\arguments{
\item{feelncOutfile}{Path to the FEELnc output file.}
}
\value{
A named character vector of noncoding transcript IDs.
}
\description{
Read FEELnc Output
}
\examples{
# Create a temporary file
tempFeelncFile <- tempfile()
write.table(
 data.frame(name=c("T1", "T2", "T3"), label=c(0, 1, 0)),
 tempFeelncFile, row.names=FALSE, col.names=TRUE, quote=FALSE, sep="\t"
)
# Run and print
noncodingFeelnc <- readFeelnc(tempFeelncFile)
print(noncodingFeelnc)
# Clean up
unlink(tempFeelncFile)
}
\keyword{internal}
