% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodPotV2.R
\name{readLncRnaMdeep}
\alias{readLncRnaMdeep}
\title{Read lncRNA_Mdeep Output}
\usage{
readLncRnaMdeep(lncRnaMdeepOutfile)
}
\arguments{
\item{lncRnaMdeepOutfile}{Path to the lncRNA_Mdeep output file.}
}
\value{
A named character vector of noncoding transcript IDs.
}
\description{
Read lncRNA_Mdeep Output
}
\examples{
# Create a temporary file
tempMdeepFile <- tempfile()
write.table(
 data.frame(V1=c("seqA", "seqB"), V2=c("noncoding", "coding")),
 tempMdeepFile, row.names=FALSE, col.names=FALSE, quote=FALSE
)
# Run and print
noncodingMdeep <- readLncRnaMdeep(tempMdeepFile)
print(noncodingMdeep)
# Clean up
unlink(tempMdeepFile)
}
\keyword{internal}
