% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodPotV2.R
\name{readPlek}
\alias{readPlek}
\title{Read PLEK Output}
\usage{
readPlek(plekOutfile)
}
\arguments{
\item{plekOutfile}{Path to the PLEK output file.}
}
\value{
A named character vector of noncoding transcript IDs.
}
\description{
Read PLEK Output
}
\examples{
# Create a temporary file
tempPlekFile <- tempfile()
write.table(
 data.frame(V1=c("Non-coding", "Coding"), V3=c(">ID_A", ">ID_B")),
 tempPlekFile, row.names=FALSE, col.names=FALSE, quote=FALSE
)
# Run and print
noncodingPlek <- readPlek(tempPlekFile)
print(noncodingPlek)
# Clean up
unlink(tempPlekFile)
}
\keyword{internal}
