% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regGTF.R
\name{refBiotypes}
\alias{refBiotypes}
\title{This function allows you to extract some statistics from reference GTF file}
\usage{
refBiotypes(refGTF, genes = F)
}
\arguments{
\item{refGTF}{is the reference GTF imported with 'importGFF()' function from 'rtracklayer' package}

\item{genes}{can have to values TRUE - extract information about genes and FALSE - extract informations about transcripts}
}
\description{
This function allows you to extract some statistics from reference GTF file
}
\examples{
refBiotypes('GTF from stringtie by importGFF()')
}
\keyword{GTF}
\keyword{lncRNA}
