% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strGTF2stat.R
\name{strGTF2stat}
\alias{strGTF2stat}
\title{A strGTF2stat function}
\usage{
strGTF2stat(stringtieGTF)
}
\arguments{
\item{stringtieGTF}{is the GTF merged by stringtie and imported with 'importGFF()' function from 'rtracklayer' package}
}
\description{
This function allows you to extract some statistics from GTF file
}
\keyword{GTF}
\keyword{lncRNA}
\keyword{stringtie}
