% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringtie2express.R
\name{strtie2expr_units}
\alias{strtie2expr_units}
\title{strtie2expr_units function}
\usage{
strtie2expr_units(strdir = "stringtie/", pattern = "gtab")
}
\arguments{
\item{strdir}{the directory with stringtie results (default is strdir = "stringtie/")}

\item{pattern}{the pattern for stringtie uotput files (default is pattern = "gtab")}
}
\description{
This function allows you to list the possible variables from stringtie output file.
}
\examples{
strtie2expr_units()
}
\keyword{expression}
\keyword{lncRNA}
