% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringtie2express.R
\name{strtie2glen}
\alias{strtie2glen}
\title{strtie2glen function}
\usage{
strtie2glen(strdir = "stringtie/", pattern = "gtab")
}
\description{
This function allows you to compute genes length from stringtie output file.
}
\examples{
strtie2glen()
}
\keyword{expression}
\keyword{lncRNA}
