% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumCombTools.R
\name{SumCombTools}
\alias{SumCombTools}
\title{SumCombTools function}
\usage{
SumCombTools(BestPat, selectedTools)
}
\arguments{
\item{BestPat}{the table with summarized data generated with SumSingleTools() function}

\item{tools}{the list of names of tools to compute the accuracy}
}
\description{
This function summarise coding potential results for all combinations of selected tools.
}
\examples{
BestPat3 <- SumAtLeast(BestPat = BestPat2, tools = selectedTools)
head(BestPat3)
}
\keyword{lncRNA}
\keyword{test}
\keyword{training}
