% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumSingleTools.R
\name{SumSingleTools}
\alias{SumSingleTools}
\title{SumSingleTools function}
\usage{
SumSingleTools(CodPot.tbl2, nc_test, cds_test)
}
\arguments{
\item{CodPot.tbl2}{the table with categorical values (0,1) where 0 means high coding potential and 1 means non-coding sequence.}

\item{nc_test}{the list of names of non-coding sequences for accuracy test (should be as the part of sequence names in the CodPot.tbl2)}

\item{cds_test}{the list of names of protein coding sequences for accuracy test (should be as the part of sequence names in the CodPot.tbl2)}
}
\description{
This function summarise coding potential results for single tools.
}
\examples{
BestPat1 <- SumSingleTools(CodPot.tbl2 = tbl2, nc_test = nc_tt$nc.test, cds_test = cds_tt$cds.test)
}
\keyword{lncRNA}
\keyword{test}
\keyword{training}
