% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.train.cds.R
\name{test.train.cds}
\alias{test.train.cds}
\title{test.train.cds function}
\usage{
test.train.cds(cds.fa, percent_train = 0.6)
}
\arguments{
\item{cds.fa}{the variable containing sequences in fasta format (from read.fasta() function from seqinr library)}

\item{percent_train}{the ratio value to divide the input set (he default is 0.6, which means that 60\\% of the sequence is to be allocated to the training set)}
}
\description{
This function allows you to split a fast format sequence file into two sets: test and train.
As a result, the function outputs a list consisting of two elements "test" and "train" containing sequence names.
}
\examples{
# cds <- read.fasta("lncRNA/Mus_musculus.GRCm39.cds.all.fa", seqtype = "DNA", as.string = T, set.attributes = F)
# cds_tt <- test.train.cds(cds.fa = cds, percent_train = 0.6)
}
