% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{TransAct}
\alias{TransAct}
\title{A TransAct function}
\usage{
TransAct(
  expr.matrix,
  cor.method = "pearson",
  rval = 0.7,
  pval = 0.05,
  lncRNA.list = NULL,
  tarRNA.list = NULL,
  full.cor.matrix.filename = NULL
)
}
\arguments{
\item{expr.matrix}{espression matrix; rownames should contains the genes or transcripts names and collumns shoud be the samples.}

\item{cor.method}{correlation metthod: "pearson" or "spearman"}

\item{rval}{the cutoff of correlation coeficient; default value is 0.7}

\item{pval}{the cutoff of estimated p-value for any interaction; default value is 0.05}

\item{lncRNA.list}{the list of lncRNA genes/transcripts. The manes shoud be a part of "expr.matrix" rownames.}

\item{tarRNA.list}{the list of target genes/transcripts. The manes shoud be a part of "expr.matrix" rownames.}
}
\description{
This function allows to estimate trans interactions between lncRNA's and potential target genes basing on the expression/coexpression levels.
This function requires the "Hmisc" library to be installed
}
\examples{
transNjJd <- TransAct(expr.matrix = em_NjJd, rval = 0.9, lncRNA.list = DELsNjJd, tarRNA.list = pcDEGsNjJd)
head(transNjJd)
TransAct()

}
\keyword{acting}
\keyword{lncRNA}
\keyword{trans}
