% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestToolComb.R
\name{validateBestToolCombInput}
\alias{validateBestToolCombInput}
\title{Validate Input for bestToolComb Function}
\usage{
validateBestToolCombInput(sumCombToolsList)
}
\arguments{
\item{sumCombToolsList}{A list expected to be generated by \code{sumCombTools()}.}
}
\value{
TRUE if validation passes, FALSE otherwise.
}
\description{
Checks if the input list is valid for the bestToolComb function.
}
\keyword{internal}
