% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestToolComb.R
\name{validateCombinationNames}
\alias{validateCombinationNames}
\title{Validate Provided Tool Combination Names}
\usage{
validateCombinationNames(combsToValidate, availableCombs)
}
\arguments{
\item{combsToValidate}{Character vector of combination names to check.}

\item{availableCombs}{Character vector of all available combination names.}
}
\value{
Character vector of validated combination names.
}
\description{
Checks if provided combination names exist in the list of available combinations.
}
\keyword{internal}
