% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodPot.R
\name{venn.CodPot}
\alias{venn.CodPot}
\title{A venn.CodPot function}
\usage{
venn.CodPot(
  CodPot,
  venncolors = c("green", "red", "blue", "yellow", "magenta", "black", "orange", "white",
    "darkgreen"),
  selmet = NULL
)
}
\arguments{
\item{CodPot}{is the table generated usig CodPot2tbl() function}

\item{selmet}{is the vector argument indicatin which methods you select to venn diagram - i.e.: c(1,1,1,1) or c(1,0,1,0).}
}
\description{
This function allows to draw Venn diagram od coding potential outputs
}
\examples{
venn.CodPot(CodPot = CodPot_table, selmet = c(1,1,0,0,1))

}
\keyword{lncRNA}
\keyword{venn.CodPot}
