% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodPot.R
\name{CodPot2tbl}
\alias{CodPot2tbl}
\title{A CodPot2tbl function}
\usage{
CodPot2tbl(
  CPC2_outfile = NULL,
  PLEK_outfile = NULL,
  FEELnc_outfile = NULL,
  CPAT_outfile = NULL,
  CPAT_cutoff = 0.364,
  CNCI_outfile = NULL,
  LncFinder_outfile = NULL,
  lncRNA_Mdeep_outfile = NULL
)
}
\description{
This function reads several outputs from coding potential programs and join all tesults in table, where 0 is coding and 1 - noncoding
}
\examples{
CodPot2tbl()

}
\keyword{coding}
\keyword{lncRNA}
\keyword{potential}
