% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringtie2express.R
\name{strtie2expr}
\alias{strtie2expr}
\title{strtie2expr function}
\usage{
strtie2expr(strdir = "stringtie/", pattern = "gtab", expunit = "FPKM")
}
\arguments{
\item{strdir}{the directory with stringtie results (default is strdir = "stringtie/")}

\item{pattern}{the pattern for stringtie uotput files (default is pattern = "gtab")}
}
\description{
This function allows you to extract expression values from the stringtie output file.
}
\examples{
strtie2expr()
FPKM_strtie <- strtie2expr()
FPKM_strtie <- strtie2expr(strdir = "stringtie/", pattern = "gtab", expunit = "FPKM")
TPM_strtie <- strtie2expr(expunit = "TPM")
}
\keyword{expression}
\keyword{lncRNA}
