% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringtie2express.R
\name{strtie2glen}
\alias{strtie2glen}
\title{strtie2glen function}
\usage{
strtie2glen(strdir = "stringtie/", pattern = "gtab")
}
\arguments{
\item{strdir}{the directory with stringtie results (default is strdir = "stringtie/")}

\item{pattern}{the pattern for stringtie uotput files (default is pattern = "gtab")}
}
\description{
This function allows you to compute genes length from stringtie output file.
}
\examples{
strtie2glen()
}
\keyword{expression}
\keyword{lncRNA}
