% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumAtLeast.R
\name{SumAtLeast}
\alias{SumAtLeast}
\title{SumAtLeast function}
\usage{
SumAtLeast(BestPat, tools)
}
\arguments{
\item{BestPat}{the table with summarized data generated with SumSingleTools() function}

\item{tools}{the list of names of tools to compute the accuracy}
}
\description{
This function summarise coding potential results for several tools counting if any sequence is predicted by at least n tools.
}
\examples{
BestPat2 <- SumAtLeast(BestPat = BestPat1, tools = selectedTools)
head(BestPat2)

}
\keyword{lncRNA}
\keyword{test}
\keyword{training}
